+++ app/models/password_reset.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, VARBINARY, DateTime
+from sqlalchemy import Column, Integer, VARBINARY, DateTime
 from sqlalchemy.orm import relationship
 from pydantic import BaseModel
 from datetime import datetime
@@ -8,8 +8,8 @@ from app.models.base import Base
 class PasswordReset(Base):
     __tablename__ = "password_resets"
     
-    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
-    user_id = Column(BigInteger, nullable=False, index=True)
+    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
+    user_id = Column(Integer, nullable=False, index=True)
     token_hash = Column(VARBINARY(32), nullable=False, unique=True)
     expires_at = Column(DateTime, nullable=False, index=True)
     used_at = Column(DateTime, nullable=True)
