+++ app/models/role_permission.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, BigInteger, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 from pydantic import BaseModel
 from datetime import datetime
@@ -8,8 +8,8 @@ from app.models.base import Base
 class RolePermission(Base):
     __tablename__ = "role_permissions"
     
-    role_id = Column(BigInteger, ForeignKey("roles.id"), primary_key=True)
-    permission_id = Column(BigInteger, ForeignKey("permissions.id"), primary_key=True)
+    role_id = Column(Integer, ForeignKey("roles.id"), primary_key=True)
+    permission_id = Column(Integer, ForeignKey("permissions.id"), primary_key=True)
     created_at = Column(DateTime, nullable=True, default=datetime.utcnow)
     
     # 關聯
